Add check

+++ backend/api/serializers.py
@@ -113,6 +113,8 @@ class RecipeCreateUpdateSerializer(serializers.ModelSerializer):
         """Создает рецепт."""
         ingredients = validated_data.pop('ingredients')
         tags = validated_data.pop('tags')
+        if len(ingredients) == 0 or len(tags)== 0:
+            raise serializers.ValidationError('Заполните все необходимые поля')
         recipe = Recipe.objects.create(**validated_data)
         for ingredient in ingredients:
             RecipeIngredient.objects.create(recipe=recipe, ingredient=ingredient['id'],
@@ -129,6 +131,8 @@ class RecipeCreateUpdateSerializer(serializers.ModelSerializer):
         instance.image = validated_data.get('image', instance.image)
         ingredients = validated_data.pop('ingredients')
         tags = validated_data.pop('tags')
+        if len(ingredients) == 0 or len(tags)== 0:
+            raise serializers.ValidationError('Заполните все необходимые поля')
         recipe_ingr = RecipeIngredient.objects.filter(recipe=instance)
         recipe_ingr.delete()
         for ingredient in ingredients:
